pyproject.toml
[project]
name = "med-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "fastmcp>=2.10.6",
    "google-adk>=1.8.0",
    "litellm>=1.74.8",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "toolbox-core>=0.4.0",
]
